Add CI - Run Tests (v4)
This makes it so that on every push (includes PRs), dependencies and tests run on GitHub Actions.

It also adds a small badge reporting the latest push/run on the `v4` branch.

Fixes #321

+++ packages/core/test/account.ts
@@ -35,6 +35,6 @@ suite("Account", () => {
         await account.unlock(password);
 
         assert.equal(bytesToHex(publicKey), bytesToHex(account.publicKey));
-        assert.equal(bytesToHex(privateKey), bytesToHex(account.privateKey));
+        assert.equal(bytesToHex(privateKey as Uint8Array), bytesToHex(account.privateKey as Uint8Array));
     });
 });
